+++ src/net/sf/freecol/common/model/SimpleCombatModel.java
@@ -94,14 +94,14 @@ public class SimpleCombatModel extends CombatModel {
         }
 
         if (lb != null) lb.add(" attacker=", attacker, " ");
-        float attackPower = getOffencePower(attacker, defender, lb);
+        double attackPower = getOffencePower(attacker, defender, lb);
         if (lb != null) lb.add(" defender=", defender, " ");
-        float defencePower = getDefencePower(attacker, defender, lb);
-        if (attackPower == 0.0f && defencePower == 0.0f) {
+        double defencePower = getDefencePower(attacker, defender, lb);
+        if (attackPower == 0.0 && defencePower == 0.0) {
             if (lb != null) lb.add(" odds=unknown");
             return new CombatOdds(CombatOdds.UNKNOWN_ODDS);
         }
-        float victory = attackPower / (attackPower + defencePower);
+        double victory = attackPower / (attackPower + defencePower);
         if (lb != null) lb.add(" odds=", victory);
         return new CombatOdds(victory);
     }
@@ -114,8 +114,8 @@ public class SimpleCombatModel extends CombatModel {
      * @return The offensive power.
      */
     @Override
-    public float getOffencePower(FreeColGameObject attacker,
-                                 FreeColGameObject defender) {
+    public double getOffencePower(FreeColGameObject attacker,
+                                  FreeColGameObject defender) {
         return getOffencePower(attacker, defender, null);
     }
 
@@ -140,10 +140,10 @@ public class SimpleCombatModel extends CombatModel {
      * @param lb An optional <code>LogBuilder</code> to log to.
      * @return The offensive power.
      */
-    private float getOffencePower(FreeColGameObject attacker,
-                                  FreeColGameObject defender,
-                                  LogBuilder lb) {
-        float result = 0.0f;
+    private double getOffencePower(FreeColGameObject attacker,
+                                   FreeColGameObject defender,
+                                   LogBuilder lb) {
+        double result = 0.0;
         if (attacker == null) {
             throw new IllegalStateException("Null attacker");
 
@@ -184,7 +184,7 @@ public class SimpleCombatModel extends CombatModel {
      * @return The defensive power.
      */
     @Override
-    public float getDefencePower(FreeColGameObject attacker,
+    public double getDefencePower(FreeColGameObject attacker,
                                  FreeColGameObject defender) {
         return getDefencePower(attacker, defender, null);
     }
@@ -197,10 +197,10 @@ public class SimpleCombatModel extends CombatModel {
      * @param lb An optional <code>LogBuilder</code> to log to.
      * @return The defensive power.
      */
-    public float getDefencePower(FreeColGameObject attacker,
+    public double getDefencePower(FreeColGameObject attacker,
                                  FreeColGameObject defender,
                                  LogBuilder lb) {
-        float result;
+        double result;
         if (combatIsDefenceMeasurement(attacker, defender)
             || combatIsAttack(attacker, defender)
             || combatIsSettlementAttack(attacker, defender)
@@ -602,7 +602,7 @@ public class SimpleCombatModel extends CombatModel {
         lb.add("Combat");
         ArrayList<CombatResult> crs = new ArrayList<>();
         CombatOdds odds = calculateCombatOdds(attacker, defender, lb);
-        float r = randomFloat(logger, "AttackResult", random);
+        double r = randomDouble(logger, "AttackResult", random);
         lb.add(" random(1.0)=", r);
         boolean great = false; // Great win or loss?
         String action;
@@ -612,7 +612,7 @@ public class SimpleCombatModel extends CombatModel {
             Unit defenderUnit = (Unit) defender;
             action = "Attack";
 
-            // For random float 0 <= r < 1.0:
+            // For random double 0 <= r < 1.0:
             // Partition this range into wins < odds.win and losses above.
             // Within the 0 <= r < odds.win range, partition the first 10%
             // to be great wins and the rest to be ordinary wins.
@@ -628,22 +628,22 @@ public class SimpleCombatModel extends CombatModel {
             //   else => great loss
             // ...and beached ships always lose.
             if (r < odds.win || defenderUnit.isBeached()) {
-                great = r < 0.1f * odds.win; // Great Win
+                great = r < 0.1 * odds.win; // Great Win
                 crs.add(CombatResult.WIN);
                 resolveAttack(attackerUnit, defenderUnit, great,
                     // Rescale to 0 <= r < 1
-                    r / (0.1f * odds.win), crs);
-            } else if (r < 0.8f * odds.win + 0.2f
+                    r / (0.1 * odds.win), crs);
+            } else if (r < 0.8 * odds.win + 0.2
                     && defenderUnit.hasAbility(Ability.EVADE_ATTACK)) {
                 crs.add(CombatResult.NO_RESULT);
                 crs.add(CombatResult.EVADE_ATTACK);
             } else {
-                great = r >= 0.1f * odds.win + 0.9f; // Great Loss
+                great = r >= 0.1 * odds.win + 0.9; // Great Loss
                 crs.add(CombatResult.LOSE);
                 resolveAttack(defenderUnit, attackerUnit, great,
                     // Rescaling to 0 <= r < 1
                     // (rearrange: 0.8 * odds.win + 0.2 <= r < 1.0)
-                    (1.25f * r - 0.25f - odds.win)/(1.0f - odds.win), crs);
+                    (1.25 * r - 0.25 - odds.win)/(1.0 - odds.win), crs);
             }
 
         } else if (combatIsBombard(attacker, defender)) {
@@ -661,9 +661,9 @@ public class SimpleCombatModel extends CombatModel {
 
                 // Great wins occur at most in 1 in 3 of successful bombards,
                 // Good defences reduce this proportion.
-                float offencePower = getOffencePower(attacker, defender);
-                float defencePower = getDefencePower(attacker, defender);
-                float diff = Math.max(3f, defencePower * 2f - offencePower);
+                double offencePower = getOffencePower(attacker, defender);
+                double defencePower = getDefencePower(attacker, defender);
+                double diff = Math.max(3.0, defencePower * 2.0 - offencePower);
                 great = r < odds.win / diff;
 
                 // Sink the defender on great wins or lack of repair
@@ -706,7 +706,7 @@ public class SimpleCombatModel extends CombatModel {
      * @param crs A list of <code>CombatResult</code>s to add to.
      */
     private void resolveAttack(Unit winner, Unit loser, boolean great,
-                               float r, List<CombatResult> crs) {
+                               double r, List<CombatResult> crs) {
         Player loserPlayer = loser.getOwner();
         Tile tile = loser.getTile();
         Player winnerPlayer = winner.getOwner();
@@ -805,7 +805,7 @@ public class SimpleCombatModel extends CombatModel {
                             crs.add(CombatResult.CAPTURE_CONVERT);
                             lose++;
                         }
-                    } else if (r >= 1.0f - winner.getBurnProbability()) {
+                    } else if (r >= 1.0 - winner.getBurnProbability()) {
                         for (IndianSettlement s
                                  : loserPlayer.getIndianSettlements()) {
                             if (s.hasMissionary(winnerPlayer)) {
